backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0
itsdangerous==2.1.2
slowapi==0.1.9

# Image Processing
pillow==10.2.0
colorthief==0.2.1

# Validation
pydantic[email]==2.5.3
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
